+++ test/DataTableSerializer.ts
@@ -34,8 +34,8 @@ describe("DataTableSerializer", function() {
 
         // TODO: add different data types
         var d1 = new DataTable("myTable");
-        d1.rows.push({ prop1: Guid.create().toString(), prop2: "value2&" }); // use xml not permitted values
-        d1.rows.push({ prop1: true, prop2: "val\tue2\n", prop3: "0001" });
+        d1.rows.push({ prop1: Guid.create().toString(), prop2: "value2&<>'\"" }); // use xml not permitted values
+        d1.rows.push({ prop1: true, prop2: "val\tue2\n", prop3: "0001", "1 <&\'>\" ":"test" }); // use xml not permitted values as property names
         d1.rows.push({ prop1: false, prop2: new Date(), prop3: 12, prop4: 12.5, prop5: "12345" });
         d1.rows.push({ prop1: "whatever", prop2: { type: "myType", value: "my value" } });
         DataTableSerializer.save(d1,fileName);